// web/src/api/generated/models/RelatedPersonFields.ts
export interface RelatedPersonFields {
    /**
     * 
     * @type {string}
     * @memberof RelatedPersonFields
     */
    avatarAttachmentId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof RelatedPersonFields
     */
    birthday?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof RelatedPersonFields
     */
    birthdayYearKnown?: boolean;
    /**
     * 
     * @type {string}
     * @memberof RelatedPersonFields
     */
    displayName: string;
    /**
     * 
     * @type {PersonRelationship}
     * @memberof RelatedPersonFields
     */
    relationship: PersonRelationship;
    /**
     * 
     * @type {boolean}
     * @memberof RelatedPersonFields
     */
    showBirthdayOnDashboard?: boolean;
    /**
     * 
     * @type {ContentVisibility}
     * @memberof RelatedPersonFields
     */
    visibility: ContentVisibility;
}